src/test/java/com/nutfreedom/mvc/converter/CategoryToCategoryCommandTest.java
package com.nutfreedom.mvc.converter;

import com.nutfreedom.mvc.command.CategoryCommand;
import com.nutfreedom.mvc.entity.Category;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class CategoryToCategoryCommandTest {
    private static final Long ID_VALUE = 1L;
    private static final String DESCRIPTION = "description";
    private CategoryToCategoryCommand converter;

    @Before
    public void setUp() {
        converter = new CategoryToCategoryCommand();
    }

    @Test
    public void convert() {
        Category category = new Category();
        category.setId(ID_VALUE);
        category.setDescription(DESCRIPTION);

        CategoryCommand categoryCommand = converter.convert(category);

        assertNotNull(categoryCommand);
        assertEquals(ID_VALUE, categoryCommand.getId());
        assertNotNull(DESCRIPTION, categoryCommand.getDescription());
    }
}
